requirements.txt
# Core dependencies
streamlit==1.32.2
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
wordcloud==1.9.3

# Kafka dependencies
kafka-python==2.0.2

# AWS S3 access
boto3==1.34.76
s3fs==2024.3.1
fsspec==2024.3.1

# Secret detection (optional for dev)
detect-secrets==1.4.0

# Spark support (optional for Databricks integration)
pyspark==3.5.1
